Cleanup old code (#40)
