--- core/source/operands/debug.cpp
@@ -1,46 +1,0 @@
-/**
- * Copyright 2020 Michel Kremer (kremi151)
- *
- * Licensed under the Apache License, Version 2.0 (the License);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "debug.h"
-
-#include <operands/instruction_context.h>
-
-#ifdef FB_DEBUG_WRITE_EXECUTION_LOG
-
-using namespace FunkyBoy;
-
-void Debug::writeExecutionToLog(uint8_t discriminator, std::ofstream &file, FunkyBoy::InstrContext &instrContext) {
-    file << discriminator << " ";
-    file << "0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (instrContext.instr & 0xff);
-    file << " B=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regB & 0xff);
-    file << " C=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regC & 0xff);
-    file << " D=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regD & 0xff);
-    file << " E=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regE & 0xff);
-    file << " H=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regH & 0xff);
-    file << " L=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regL & 0xff);
-    file << " A=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regA & *instrContext.regA);
-    file << " F=0x" << std::uppercase << std::setfill('0') << std::setw(2) << std::hex << (*instrContext.regF & 0xff);
-    file << " PC=0x" << std::uppercase << std::setfill('0') << std::setw(4) << std::hex << (instrContext.progCounter - 1);
-    file << " SP=0x" << std::uppercase << std::setfill('0') << std::setw(4) << std::hex << instrContext.stackPointer;
-    file << std::endl;
-}
-
-void Debug::writeInterruptToLog(uint16_t interrupt, std::ofstream &file) {
-    file << "Int 0x" << std::uppercase << std::setfill('0') << std::setw(4) << std::hex << interrupt;
-    file << std::endl;
-}
-
-#endif
