--- core/source/operands/instructions.h
@@ -1,25 +1,0 @@
-/**
- * Copyright 2020 Michel Kremer (kremi151)
- *
- * Licensed under the Apache License, Version 2.0 (the License);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef FB_CORE_OPERANDS_H
-#define FB_CORE_OPERANDS_H
-
-#include <operands/instruction_context.h>
-#include <operands/reads.h>
-#include <operands/writes.h>
-#include <operands/conditions.h>
-
-#endif //FB_CORE_OPERANDS_H
